Added a placeholder for a plugin loading infrastructure.
This is based on a discussion between Dave MacMahon and Alec Rust
on the CASPER mailing list. A "load-plugin" command is added to the
raw katcp mode of tcpborphserver. This command will load a dynamic
library and add new commands to the server.

Currently it just prints the plugin name (hence place-holder).