Fleshed out the plugin instrasture.
Each plugin must contain an instance of struct PLUGIN (defined
in plugin.h) called KATCP_PLUGIN which itself contains an array
of structs PLUGIN_CMD which define new commands for katcp. The
dynamic loader looks for these and loads in commands appropriately.

Also, changed the load command to "?plugin-load" and added a
"?plugin-list" command which lists currently loaded plugins. There's
also "?plugin-unload" which should remove loaded plugins but this
remains to be written.