[katcp] Force katcp's BUILD timestamp to be UTC
Also include "UTC" timezone designator at the end of the string.  This
adds whitespace in the BUILD string (e.g. "2013-04-03T05:34:59 UTC")
which complicates the passing of the BUILD string since it was wrapped
in double quotes that were escaped from the shell and passed through to
the compiler (to make the macro expand to a double quated string).  This
is addressed by NOT escaping the quotes from the shell and then
stringifying the macro text using a new preprocessor macro named
STRINGIFY (defined in katcp/katcp.h, see comments there for more info).